# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ProjectPulse API
  version: 1.0.0
  description: Token-authenticated API shared by the ProjectPulse admin web and member mobile apps.
servers:
  - url: http://localhost:8000/api
    description: Local Docker or Laravel server
...
paths:
...
  /tasks/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Tasks]
      security: [{ bearerAuth: [] }]
      summary: Get an authorized task
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "403": { $ref: "#/components/responses/Forbidden" }
    patch:
      tags: [Tasks]
      security: [{ bearerAuth: [] }]
      summary: Update task fields (admin)
      requestBody:
        $ref: "#/components/requestBodies/Task"
      responses:
        "200": { $ref: "#/components/responses/Success" }
    delete:
      tags: [Tasks]
      security: [{ bearerAuth: [] }]
      summary: Soft-delete a task (admin)
      responses:
        "200": { $ref: "#/components/responses/Success" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: integer } }
...
  requestBodies:
...
    Task:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/TaskInput" }
...
  responses:
    Success:
      description: Successful envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SuccessEnvelope" }
...
    Forbidden: { description: Insufficient permission, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
...
  schemas:
...
    TaskStatus:
      type: string
      enum: [todo, in_progress, review, done]
    TaskInput:
      type: object
      required: [title, category, priority]
      properties:
        title: { type: string, maxLength: 255 }
        description: { type: [string, "null"], maxLength: 5000 }
        category: { type: string, enum: [frontend, backend, design, qa, devops, management, other] }
        assignee_id: { type: [integer, "null"] }
        priority: { type: string, enum: [low, medium, high, urgent] }
        status: { $ref: "#/components/schemas/TaskStatus" }
        estimated_hours: { type: [number, "null"], minimum: 0, maximum: 1000 }
        start_date: { type: [string, "null"], format: date }
        deadline: { type: [string, "null"], format: date }
        source: { type: string, enum: [manual, ai], default: manual }
...
    SuccessEnvelope:
      type: object
      required: [success, message, data]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data: {}
        meta: { type: [object, "null"] }
...
    ErrorEnvelope:
      type: object
      required: [success, message, error]
      properties:
        success: { type: boolean, const: false }
        message: { type: string }
        error:
          type: object
          required: [code]
          properties:
            code:
              type: string
              enum: [UNAUTHENTICATED, FORBIDDEN, RESOURCE_NOT_FOUND, VALIDATION_ERROR, INVALID_STATUS_TRANSITION, AI_PROVIDER_UNAVAILABLE, AI_REQUEST_TIMEOUT, AI_INVALID_RESPONSE, RATE_LIMIT_EXCEEDED, INTERNAL_SERVER_ERROR]
            details: {}